backend/requirements.txt
# ── Core API server ───────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
requests>=2.31.0
python-dotenv>=1.0.0


# ── Database & Auth ───────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# ── Image processing ──────────────────────────────────────────────────────────
numpy>=1.26.0
Pillow>=10.0.0
opencv-python-headless>=4.9.0

# ── Vector search ─────────────────────────────────────────────────────────────
faiss-cpu>=1.7.0

# ── AI models ─────────────────────────────────────────────────────────────────
# PyTorch — install from PyTorch index for Python 3.13 CPU support:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0

# CLIP — install after torch
git+https://github.com/openai/CLIP.git

# YOLOv8 (optional)
ultralytics>=8.1.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
tqdm>=4.66.0
ftfy>=6.1.0
regex>=2023.12.0

